llama_cpp_agent/requirements.txt
llama-cpp-agent
